Fit the model using Sequential Monte Carlo (SMC).

Fit the model using Sequential Monte Carlo (SMC).

fitSpectraSMC(wl, spc, peakWL, lPriors, conc = rep(1, nrow(spc)),
  npart = 10000, rate = 0.9, minESS = npart/2, destDir = NA)

Arguments

wl
Vector of nwl wavenumbers at which the spetra are observed.

spc
n_y * nwl Matrix of observed Raman spectra.

peakWL
Vector of locations for each peak (cm^-1)

lPriors
List of hyperparameters for the prior distributions.

conc
Vector of n_y nanomolar (nM) dye concentrations for each observation.

npart
number of SMC particles to use for the importance sampling distribution.

rate
the target rate of reduction in the effective sample size (ESS).

minESS
minimum effective sample size, below which the particles are resampled.

destDir
destination directory to save intermediate results (for long-running computations)

Value

a List containing weighted parameter values, known as particles:

weights

Vector of importance weights for each particle.

beta

npart * npeaks Matrix of regression coefficients for the amplitudes.

scale

npart * npeaks Matrix of scale parameters.

sigma

Vector of npart standard deviations.

alpha

bl.knots * n_y * npart Array of spline coefficients for the baseline.

basis

A dense nwl * bl.knots Matrix containing the values of the basis functions.

expFn

npart * nwl Matrix containing the spectral signature.

ess

Vector containing the effective sample size (ESS) at each SMC iteration.

logEvidence

Vector containing the logarithm of the model evidence (marginal likelihood).

accept

Vector containing the Metropolis-Hastings acceptance rate at each SMC iteration.

sd.mh

niter * 2npeaks Matrix of random walk MH bandwidths at each SMC iteration..

References

Chopin (2002) "A Sequential Particle Filter Method for Static Models," Biometrika 89(3): 539--551, DOI: 10.1093/biomet/89.3.539

Examples

wavenumbers <- seq(200,600,by=10)
spectra <- matrix(nrow=1, ncol=length(wavenumbers))
peakLocations <- c(300,500)
peakAmplitude <- c(10000,4000)
peakScale <- c(10, 15)
signature <- weightedLorentzian(peakLocations, peakScale, peakAmplitude, wavenumbers)
baseline <- 1000*cos(wavenumbers/200) + 2*wavenumbers
spectra[1,] <- signature + baseline + rnorm(length(wavenumbers),0,200)
lPriors <- list(scale.mu=log(11.6) - (0.4^2)/2, scale.sd=0.4, bl.smooth=10^11, bl.knots=20,
                beta.mu=5000, beta.sd=5000, noise.sd=200, noise.nu=4)
result <- fitSpectraSMC(wavenumbers, spectra, peakLocations, lPriors, npart=500)

#> [1] "SMC with 1 observations at 1 unique concentrations, 500 particles, and 41 wavenumbers."
#> [1] "Step 0: computing 20 B-spline basis functions took 0.139999999999873 sec."
#> [1] "Step 1: initialization for 2 peaks took 0.012000000000171 sec."
#> [1] "Observation 1 20180218-194903"
#> [1] "Baselines took 0.0139999999996689 sec"
#> previous ESS 500 (target: 450 for observation 1 of 1; wavenumber 1 of 41)
#> Required 1 iterations to reduce ESS from 500 to 187.555
#> [1] "reweighting took 0.00100000000020373 sec for ESS 187.555215321432"
#> [1] "*** Resampling with 199 unique indices took 0.000999999999748979 sec ***"
#> [1] "MH acceptance rate 0.602 took 0.0170000000002801 sec."
#> previous ESS 500 (target: 450 for observation 1 of 1; wavenumber 2 of 41)
#> Required 1 iterations to reduce ESS from 500 to 306.147
#> [1] "reweighting took 0.00100000000020373 sec for ESS 306.147431594285"
#> [1] "MH acceptance rate 0.36 took 0.0219999999999345 sec."
#> previous ESS 306.147 (target: 275.533 for observation 1 of 1; wavenumber 3 of 41)
#> Required 1 iterations to reduce ESS from 306.147 to 134.676
#> [1] "reweighting took 0.00100000000020373 sec for ESS 134.675936435955"
#> [1] "*** Resampling with 194 unique indices took 0.00199999999995271 sec ***"
#> [1] "MH acceptance rate 0.488 took 0.0139999999996689 sec."
#> previous ESS 500 (target: 450 for observation 1 of 1; wavenumber 4 of 41)
#> Required 1 iterations to reduce ESS from 500 to 415.542
#> [1] "reweighting took 0.00100000000020373 sec for ESS 415.542396619262"
#> [1] "MH acceptance rate 0.294 took 0.0169999999998254 sec."
#> previous ESS 415.542 (target: 373.988 for observation 1 of 1; wavenumber 5 of 41)
#> Required 1 iterations to reduce ESS from 415.542 to 292.729
#> [1] "reweighting took 0.00100000000020373 sec for ESS 292.728581609419"
#> [1] "MH acceptance rate 0.248 took 0.0149999999998727 sec."
#> previous ESS 292.729 (target: 263.456 for observation 1 of 1; wavenumber 6 of 41)
#> Required 1 iterations to reduce ESS from 292.729 to 193.25
#> [1] "reweighting took 0.00100000000020373 sec for ESS 193.249703401955"
#> [1] "*** Resampling with 202 unique indices took 0.00199999999995271 sec ***"
#> [1] "MH acceptance rate 0.446 took 0.0239999999998872 sec."
#> previous ESS 500 (target: 450 for observation 1 of 1; wavenumber 7 of 41)
#> Required 1 iterations to reduce ESS from 500 to 342.352
#> [1] "reweighting took 0.00100000000020373 sec for ESS 342.351637301925"
#> [1] "MH acceptance rate 0.26 took 0.0150000000003274 sec."
#> previous ESS 342.352 (target: 308.116 for observation 1 of 1; wavenumber 8 of 41)
#> Required 1 iterations to reduce ESS from 342.352 to 271.088
#> [1] "reweighting took 0.000999999999748979 sec for ESS 271.087656497025"
#> [1] "MH acceptance rate 0.24 took 0.0199999999999818 sec."
#> previous ESS 271.088 (target: 243.979 for observation 1 of 1; wavenumber 9 of 41)
#> Required 1 iterations to reduce ESS from 271.088 to 193.266
#> [1] "reweighting took 0.000999999999748979 sec for ESS 193.265849324313"
#> [1] "*** Resampling with 210 unique indices took 0.00199999999995271 sec ***"
#> [1] "MH acceptance rate 0.338 took 0.0270000000000437 sec."
#> previous ESS 500 (target: 450 for observation 1 of 1; wavenumber 10 of 41)
#> Required 1 iterations to reduce ESS from 500 to 408.776
#> [1] "reweighting took 0.00199999999995271 sec for ESS 408.775799915626"
#> [1] "MH acceptance rate 0.214 took 0.0180000000000291 sec."
#> previous ESS 408.776 (target: 367.898 for observation 1 of 1; wavenumber 11 of 41)
#> Required 1 iterations to reduce ESS from 408.776 to 262.953
#> [1] "reweighting took 0.00100000000020373 sec for ESS 262.953039045606"
#> [1] "MH acceptance rate 0.204 took 0.0169999999998254 sec."
#> previous ESS 262.953 (target: 236.658 for observation 1 of 1; wavenumber 12 of 41)
#> Required 1 iterations to reduce ESS from 262.953 to 199.763
#> [1] "reweighting took 0.00100000000020373 sec for ESS 199.762953065699"
#> [1] "*** Resampling with 204 unique indices took 0.00199999999995271 sec ***"
#> [1] "MH acceptance rate 0.338 took 0.0169999999998254 sec."
#> previous ESS 500 (target: 450 for observation 1 of 1; wavenumber 13 of 41)
#> Required 1 iterations to reduce ESS from 500 to 413.487
#> [1] "reweighting took 0.000999999999748979 sec for ESS 413.486693694267"
#> [1] "MH acceptance rate 0.232 took 0.0220000000003893 sec."
#> previous ESS 413.487 (target: 372.138 for observation 1 of 1; wavenumber 14 of 41)
#> Required 1 iterations to reduce ESS from 413.487 to 352.516
#> [1] "reweighting took 0.000999999999748979 sec for ESS 352.515872812085"
#> [1] "MH acceptance rate 0.196 took 0.0149999999998727 sec."
#> previous ESS 352.516 (target: 317.264 for observation 1 of 1; wavenumber 15 of 41)
#> Required 1 iterations to reduce ESS from 352.516 to 230.889
#> [1] "reweighting took 0.00100000000020373 sec for ESS 230.888689188085"
#> [1] "*** Resampling with 209 unique indices took 0.00199999999995271 sec ***"
#> [1] "MH acceptance rate 0.352 took 0.0160000000000764 sec."
#> previous ESS 500 (target: 450 for observation 1 of 1; wavenumber 16 of 41)
#> Required 1 iterations to reduce ESS from 500 to 367.979
#> [1] "reweighting took 0.000999999999748979 sec for ESS 367.978619768101"
#> [1] "MH acceptance rate 0.202 took 0.0180000000000291 sec."
#> previous ESS 367.979 (target: 331.181 for observation 1 of 1; wavenumber 17 of 41)
#> Required 1 iterations to reduce ESS from 367.979 to 323.046
#> [1] "reweighting took 0.00100000000020373 sec for ESS 323.045840853913"
#> [1] "MH acceptance rate 0.192 took 0.0210000000001855 sec."
#> previous ESS 323.046 (target: 290.741 for observation 1 of 1; wavenumber 18 of 41)
#> Required 1 iterations to reduce ESS from 323.046 to 282.403
#> [1] "reweighting took 0.00199999999995271 sec for ESS 282.402887004533"
#> [1] "MH acceptance rate 0.21 took 0.0299999999997453 sec."
#> previous ESS 282.403 (target: 254.163 for observation 1 of 1; wavenumber 19 of 41)
#> Required 1 iterations to reduce ESS from 282.403 to 194.653
#> [1] "reweighting took 0.00100000000020373 sec for ESS 194.653157042414"
#> [1] "*** Resampling with 195 unique indices took 0.00199999999995271 sec ***"
#> [1] "MH acceptance rate 0.35 took 0.0160000000000764 sec."
#> previous ESS 500 (target: 450 for observation 1 of 1; wavenumber 20 of 41)
#> Required 2 iterations to reduce ESS from 500 to 406.837
#> [1] "reweighting took 0.00199999999995271 sec for ESS 406.837453790431"
#> [1] "MH acceptance rate 0.276 took 0.0250000000000909 sec."
#> previous ESS 406.837 (target: 366.154 for observation 1 of 1; wavenumber 22 of 41)
#> Required 1 iterations to reduce ESS from 406.837 to 356.754
#> [1] "reweighting took 0.00100000000020373 sec for ESS 356.754379956719"
#> [1] "MH acceptance rate 0.22 took 0.0160000000000764 sec."
#> previous ESS 356.754 (target: 321.079 for observation 1 of 1; wavenumber 23 of 41)
#> Required 1 iterations to reduce ESS from 356.754 to 182.375
#> [1] "reweighting took 0.000999999999748979 sec for ESS 182.375478933494"
#> [1] "*** Resampling with 186 unique indices took 0.00199999999995271 sec ***"
#> [1] "MH acceptance rate 0.352 took 0.0160000000000764 sec."
#> previous ESS 500 (target: 450 for observation 1 of 1; wavenumber 24 of 41)
#> Required 2 iterations to reduce ESS from 500 to 430.613
#> [1] "reweighting took 0.00100000000020373 sec for ESS 430.61307428268"
#> [1] "MH acceptance rate 0.266 took 0.0229999999996835 sec."
#> previous ESS 430.613 (target: 387.552 for observation 1 of 1; wavenumber 26 of 41)
#> Required 2 iterations to reduce ESS from 430.613 to 348.89
#> [1] "reweighting took 0.000999999999748979 sec for ESS 348.890337112567"
#> [1] "MH acceptance rate 0.236 took 0.0220000000003893 sec."
#> previous ESS 348.89 (target: 314.001 for observation 1 of 1; wavenumber 28 of 41)
#> Required 1 iterations to reduce ESS from 348.89 to 312.643
#> [1] "reweighting took 0.00199999999995271 sec for ESS 312.642803677038"
#> [1] "MH acceptance rate 0.264 took 0.0309999999999491 sec."
#> previous ESS 312.643 (target: 281.379 for observation 1 of 1; wavenumber 29 of 41)
#> Required 1 iterations to reduce ESS from 312.643 to 250.392
#> [1] "reweighting took 0.00100000000020373 sec for ESS 250.392338711285"
#> [1] "MH acceptance rate 0.276 took 0.0209999999997308 sec."
#> previous ESS 250.392 (target: 225.353 for observation 1 of 1; wavenumber 30 of 41)
#> Required 2 iterations to reduce ESS from 250.392 to 224.368
#> [1] "reweighting took 0.00199999999995271 sec for ESS 224.367659133519"
#> [1] "*** Resampling with 208 unique indices took 0.00200000000040745 sec ***"
#> [1] "MH acceptance rate 0.434 took 0.0139999999996689 sec."
#> previous ESS 500 (target: 450 for observation 1 of 1; wavenumber 32 of 41)
#> Required 1 iterations to reduce ESS from 500 to 405.188
#> [1] "reweighting took 0.00199999999995271 sec for ESS 405.187966794126"
#> [1] "MH acceptance rate 0.276 took 0.0150000000003274 sec."
#> previous ESS 405.188 (target: 364.669 for observation 1 of 1; wavenumber 33 of 41)
#> Required 1 iterations to reduce ESS from 405.188 to 327.583
#> [1] "reweighting took 0.00100000000020373 sec for ESS 327.583126331672"
#> [1] "MH acceptance rate 0.28 took 0.0250000000000909 sec."
#> previous ESS 327.583 (target: 294.825 for observation 1 of 1; wavenumber 34 of 41)
#> Required 2 iterations to reduce ESS from 327.583 to 269.793
#> [1] "reweighting took 0.000999999999748979 sec for ESS 269.79308809383"
#> [1] "MH acceptance rate 0.306 took 0.0139999999996689 sec."
#> previous ESS 269.793 (target: 242.814 for observation 1 of 1; wavenumber 36 of 41)
#> Required 2 iterations to reduce ESS from 269.793 to 146.857
#> [1] "reweighting took 0.00100000000020373 sec for ESS 146.857177215163"
#> [1] "*** Resampling with 156 unique indices took 0.00199999999995271 sec ***"
#> [1] "MH acceptance rate 0.5 took 0.0169999999998254 sec."
#> previous ESS 500 (target: 450 for observation 1 of 1; wavenumber 38 of 41)
#> Required 2 iterations to reduce ESS from 500 to 390.185
#> [1] "reweighting took 0.00199999999995271 sec for ESS 390.184539706839"
#> [1] "MH acceptance rate 0.312 took 0.0160000000000764 sec."
#> previous ESS 390.185 (target: 351.166 for observation 1 of 1; wavenumber 40 of 41)
#> Required 2 iterations to reduce ESS from 390.185 to 309.66
#> [1] "reweighting took 0.00199999999995271 sec for ESS 309.660054197453"
#> [1] "MH acceptance rate 0.336 took 0.0270000000000437 sec."
#> [1] "SMC iter. 1 took 0.993999999999687 sec."