Fit the model with Voigt peaks using Sequential Monte Carlo (SMC).

Fit the model with Voigt peaks using Sequential Monte Carlo (SMC).

fitVoigtPeaksSMC(wl, spc, lPriors, conc = rep(1, nrow(spc)), npart = 10000,
  rate = 0.9, mcAR = 0.23, mcSteps = 10, minESS = npart/2,
  destDir = NA)

Arguments

wl
Vector of nwl wavenumbers at which the spetra are observed.

spc
n_y * nwl Matrix of observed Raman spectra.

lPriors
List of hyperparameters for the prior distributions.

conc
Vector of n_y nanomolar (nM) dye concentrations for each observation.

npart
number of SMC particles to use for the importance sampling distribution.

rate
the target rate of reduction in the effective sample size (ESS).

mcAR
target acceptance rate for the MCMC kernel

mcSteps
number of iterations of the MCMC kernel

minESS
minimum effective sample size, below which the particles are resampled.

destDir
destination directory to save intermediate results (for long-running computations)

Examples

wavenumbers <- seq(200,600,by=10)
spectra <- matrix(nrow=1, ncol=length(wavenumbers))
peakLocations <- c(300,500)
peakAmplitude <- c(10000,4000)
peakScale <- c(10, 15)
signature <- weightedLorentzian(peakLocations, peakScale, peakAmplitude, wavenumbers)
baseline <- 1000*cos(wavenumbers/200) + 2*wavenumbers
spectra[1,] <- signature + baseline + rnorm(length(wavenumbers),0,200)
lPriors <- list(scaG.mu=log(11.6) - (0.4^2)/2, scaG.sd=0.4, scaL.mu=log(11.6) - (0.4^2)/2,
   scaL.sd=0.4, bl.smooth=5, bl.knots=20, loc.mu=peakLocations, loc.sd=c(5,5),
   beta.mu=c(5000,5000), beta.sd=c(5000,5000), noise.sd=200, noise.nu=4)
result <- fitVoigtPeaksSMC(wavenumbers, spectra, lPriors, npart=50, mcSteps=1)

#> [1] "SMC with 1 observations at 1 unique concentrations, 50 particles, and 41 wavenumbers."
#> [1] "Step 0: computing 24 B-spline basis functions (r=21.0526315789474) took 0.0259999999998399sec."
#> [1] "Mean noise parameter sigma is now 241.132496036009"
#> [1] "Mean spline penalty lambda is now 5"
#> [1] "Step 1: initialization for 2 Voigt peaks took 0.0210000000001855 sec."
#> [1] "Reweighting took 0.00300000000015643sec. for ESS 44.8318293761618 with new kappa 0.02734375."
#> [1] "Temp ESS is 44.8318293761618."
#> [1] "Iteration 2 took 0.0140000000001237sec. for 1 MCMC loops (acceptance rate 0.32)"
#> [1] "Reweighting took 0.00299999999970169sec. for ESS 40.3742732560537 with new kappa 0.0387420654296875."
#> [1] "Temp ESS is 40.3742732560537."
#> [1] "Iteration 3 took 0.0119999999997162sec. for 1 MCMC loops (acceptance rate 0.14)"
#> [1] "Reweighting took 0.00199999999995271sec. for ESS 36.6888817812778 with new kappa 0.0462518930435181."
#> [1] "Temp ESS is 36.6888817812778."
#> [1] "Iteration 4 took 0.0109999999999673sec. for 1 MCMC loops (acceptance rate 0.34)"
#> [1] "Reweighting took 0.00199999999995271sec. for ESS 32.8959477747114 with new kappa 0.0537030501291156."
#> [1] "Temp ESS is 32.8959477747114."
#> [1] "Iteration 5 took 0.0109999999999673sec. for 1 MCMC loops (acceptance rate 0.18)"
#> [1] "Reweighting took 0.00399999999990541sec. for ESS 29.1246914932203 with new kappa 0.0610959950499819."
#> [1] "Temp ESS is 29.1246914932203."
#> [1] "Iteration 6 took 0.0199999999999818sec. for 1 MCMC loops (acceptance rate 0.3)"
#> [1] "Reweighting took 0.00199999999995271sec. for ESS 25.5593461751357 with new kappa 0.0684311825886539."
#> [1] "Temp ESS is 25.5593461751357."
#> [1] "Iteration 7 took 0.0109999999999673sec. for 1 MCMC loops (acceptance rate 0.2)"
#> [1] "Reweighting took 0.00199999999995271sec. for ESS 22.3693618557722 with new kappa 0.07570906397468."
#> [1] "*** Resampling with 25 unique indices took 0 sec ***"
#> [1] "Temp ESS is 22.3214285714286."
#> [1] "Iteration 8 took 0.0169999999998254sec. for 1 MCMC loops (acceptance rate 0.36)"
#> [1] "Reweighting took 0.0170000000002801sec. for ESS 19.7452312235475 with new kappa 0.0901511098500757."
#> [1] "*** Resampling with 22 unique indices took 0 sec ***"
#> [1] "Temp ESS is 15.8227848101266."
#> [1] "Iteration 9 took 0.0280000000002474sec. for 1 MCMC loops (acceptance rate 0.22)"
#> [1] "Reweighting took 0.00199999999995271sec. for ESS 14.4746281281974 with new kappa 0.0972593043043719."
#> [1] "*** Resampling with 21 unique indices took 0 sec ***"
#> [1] "Temp ESS is 15.0602409638554."
#> [1] "Iteration 10 took 0.0109999999999673sec. for 1 MCMC loops (acceptance rate 0.22)"
#> [1] "Reweighting took 0.000999999999748979sec. for ESS 12.890598850848 with new kappa 0.111364627674616."
#> [1] "*** Resampling with 21 unique indices took 0 sec ***"
#> [1] "Temp ESS is 17.8571428571429."
#> [1] "Iteration 11 took 0.00999999999976353sec. for 1 MCMC loops (acceptance rate 0.26)"
#> [1] "Reweighting took 0.00199999999995271sec. for ESS 15.1063648201996 with new kappa 0.1252495553672."
#> [1] "*** Resampling with 25 unique indices took 0 sec ***"
#> [1] "Temp ESS is 21.1864406779661."
#> [1] "Iteration 12 took 0.0170000000002801sec. for 1 MCMC loops (acceptance rate 0.2)"
#> [1] "Reweighting took 0.00199999999995271sec. for ESS 18.2046518022691 with new kappa 0.138917531064588."
#> [1] "*** Resampling with 22 unique indices took 0 sec ***"
#> [1] "Temp ESS is 16.2337662337662."
#> [1] "Iteration 13 took 0.00999999999976353sec. for 1 MCMC loops (acceptance rate 0.16)"
#> [1] "Reweighting took 0.00199999999995271sec. for ESS 13.6841222562704 with new kappa 0.152371944641704."
#> [1] "*** Resampling with 18 unique indices took 0 sec ***"
#> [1] "Temp ESS is 12.7551020408163."
#> [1] "Iteration 14 took 0.0109999999999673sec. for 1 MCMC loops (acceptance rate 0.16)"
#> [1] "Reweighting took 0.00199999999995271sec. for ESS 11.0554122400758 with new kappa 0.165616133006677."
#> [1] "*** Resampling with 19 unique indices took 0 sec ***"
#> [1] "Temp ESS is 16.8918918918919."
#> [1] "Iteration 15 took 0.0109999999999673sec. for 1 MCMC loops (acceptance rate 0.34)"
#> [1] "Reweighting took 0.00100000000020373sec. for ESS 15.2688985444359 with new kappa 0.178653380928448."
#> [1] "*** Resampling with 25 unique indices took 0 sec ***"
#> [1] "Temp ESS is 22.3214285714286."
#> [1] "Iteration 16 took 0.0180000000000291sec. for 1 MCMC loops (acceptance rate 0.18)"
#> [1] "Reweighting took 0.00199999999995271sec. for ESS 19.5486292021759 with new kappa 0.204320462774434."
#> [1] "*** Resampling with 24 unique indices took 0 sec ***"
#> [1] "Temp ESS is 23.5849056603774."
#> [1] "Iteration 17 took 0.0109999999999673sec. for 1 MCMC loops (acceptance rate 0.22)"
#> [1] "Reweighting took 0.00199999999995271sec. for ESS 20.2755522990099 with new kappa 0.216752955543583."
#> [1] "*** Resampling with 24 unique indices took 0 sec ***"
#> [1] "Temp ESS is 20.8333333333333."
#> [1] "Iteration 18 took 0.0109999999999673sec. for 1 MCMC loops (acceptance rate 0.22)"
#> [1] "Reweighting took 0.00300000000015643sec. for ESS 18.9803356155094 with new kappa 0.222872073078399."
#> [1] "*** Resampling with 19 unique indices took 0 sec ***"
#> [1] "Temp ESS is 14.2045454545455."
#> [1] "Iteration 19 took 0.012000000000171sec. for 1 MCMC loops (acceptance rate 0.26)"
#> [1] "Reweighting took 0.00199999999995271sec. for ESS 12.0578211116738 with new kappa 0.235014696936549."
#> [1] "*** Resampling with 15 unique indices took 0 sec ***"
#> [1] "Temp ESS is 18.3823529411765."
#> [1] "Iteration 20 took 0.0109999999999673sec. for 1 MCMC loops (acceptance rate 0.44)"
#> [1] "Reweighting took 0.0169999999998254sec. for ESS 15.5766935542888 with new kappa 0.246967592296915."
#> [1] "*** Resampling with 26 unique indices took 0 sec ***"
#> [1] "Temp ESS is 12.3762376237624."
#> [1] "Iteration 21 took 0.0289999999999964sec. for 1 MCMC loops (acceptance rate 0.18)"
#> [1] "Reweighting took 0.00100000000020373sec. for ESS 11.2111206604653 with new kappa 0.252850657982096."
#> [1] "*** Resampling with 19 unique indices took 0 sec ***"
#> [1] "Temp ESS is 16.2337662337662."
#> [1] "Iteration 22 took 0.0100000000002183sec. for 1 MCMC loops (acceptance rate 0.24)"
#> [1] "Reweighting took 0.000999999999748979sec. for ESS 13.8017366924448 with new kappa 0.264524866451125."
#> [1] "*** Resampling with 20 unique indices took 0 sec ***"
#> [1] "Temp ESS is 16.2337662337662."
#> [1] "Iteration 23 took 0.0109999999999673sec. for 1 MCMC loops (acceptance rate 0.18)"
#> [1] "Reweighting took 0.00100000000020373sec. for ESS 14.4836932982092 with new kappa 0.276016665412827."
#> [1] "*** Resampling with 19 unique indices took 0 sec ***"
#> [1] "Temp ESS is 21.1864406779661."
#> [1] "Iteration 24 took 0.0109999999999673sec. for 1 MCMC loops (acceptance rate 0.36)"
#> [1] "Reweighting took 0.00199999999995271sec. for ESS 18.2987582682038 with new kappa 0.298641144618676."
#> [1] "*** Resampling with 23 unique indices took 0 sec ***"
#> [1] "Temp ESS is 15.4320987654321."
#> [1] "Iteration 25 took 0.0109999999999673sec. for 1 MCMC loops (acceptance rate 0.18)"
#> [1] "Reweighting took 0.000999999999748979sec. for ESS 13.6459016699548 with new kappa 0.320558608849342."
#> [1] "*** Resampling with 18 unique indices took 0 sec ***"
#> [1] "Temp ESS is 11.9047619047619."
#> [1] "Iteration 26 took 0.00900000000001455sec. for 1 MCMC loops (acceptance rate 0.1)"
#> [1] "Reweighting took 0.00100000000020373sec. for ESS 10.2981322840487 with new kappa 0.363023695796258."
#> [1] "*** Resampling with 16 unique indices took 0 sec ***"
#> [1] "Temp ESS is 13.8888888888889."
#> [1] "Iteration 27 took 0.00900000000001455sec. for 1 MCMC loops (acceptance rate 0.2)"
#> [1] "Reweighting took 0.00100000000020373sec. for ESS 12.2178206308382 with new kappa 0.402834714808992."
#> [1] "*** Resampling with 20 unique indices took 0 sec ***"
#> [1] "Temp ESS is 10.9649122807018."
#> [1] "Iteration 28 took 0.0109999999999673sec. for 1 MCMC loops (acceptance rate 0.14)"
#> [1] "Reweighting took 0.00100000000020373sec. for ESS 9.16606571173984 with new kappa 0.44015754513343."
#> [1] "*** Resampling with 17 unique indices took 0 sec ***"
#> [1] "Temp ESS is 9.92063492063492."
#> [1] "Iteration 29 took 0.0109999999999673sec. for 1 MCMC loops (acceptance rate 0.16)"
#> [1] "Reweighting took 0.00100000000020373sec. for ESS 7.98806687927925 with new kappa 0.510137851991751."
#> [1] "*** Resampling with 17 unique indices took 0 sec ***"
#> [1] "Temp ESS is 17.3611111111111."
#> [1] "Iteration 30 took 0.0109999999999673sec. for 1 MCMC loops (acceptance rate 0.48)"
#> [1] "Reweighting took 0.000999999999748979sec. for ESS 15.2361910916581 with new kappa 0.755068925995876."
#> [1] "*** Resampling with 22 unique indices took 0 sec ***"
#> [1] "Temp ESS is 14.8809523809524."
#> [1] "Iteration 31 took 0.00999999999976353sec. for 1 MCMC loops (acceptance rate 0.14)"
#> [1] "Reweighting took 0.00100000000020373sec. for ESS 13.6112884466131 with new kappa 0.785685310246391."
#> [1] "*** Resampling with 15 unique indices took 0 sec ***"
#> [1] "Temp ESS is 12.0192307692308."
#> [1] "Iteration 32 took 0.0100000000002183sec. for 1 MCMC loops (acceptance rate 0.2)"
#> [1] "Reweighting took 0.0160000000000764sec. for ESS 10.4971478116842 with new kappa 0.839263982684793."
#> [1] "*** Resampling with 15 unique indices took 0 sec ***"
#> [1] "Temp ESS is 21.9298245614035."
#> [1] "Iteration 33 took 0.0340000000001055sec. for 1 MCMC loops (acceptance rate 0.46)"
#> [1] "Reweighting took 0.00100000000020373sec. for ESS 19.6218391637419 with new kappa 0.919631991342397."
#> [1] "*** Resampling with 23 unique indices took 0 sec ***"
#> [1] "Temp ESS is 16.8918918918919."
#> [1] "Iteration 34 took 0.01299999999992sec. for 1 MCMC loops (acceptance rate 0.24)"
#> [1] "Reweighting took 0sec. for ESS 15.3353925711179 with new kappa 1."
#> [1] "*** Resampling with 20 unique indices took 0 sec ***"
#> [1] "Temp ESS is 16.025641025641."
#> [1] "Iteration 35 took 0.0100000000002183sec. for 1 MCMC loops (acceptance rate 0.28)"