Compute the spectral signature using Voigt peaks.

Calculates the value of the pseudo-Voigt broadening function at the given wavenumbers, given the parameters of the peaks. This function is thread-safe.

mixedVoigt(location, scale_G, scale_L, amplitude, wavenum)

Arguments

location
Vector of location parameters of the peaks (\(cm^{-1}\))

scale_G
Vector of standard deviations \(\sigma_j\) of the Gaussian components.

scale_L
Vector of scale parameters \(\phi_j\) of the Lorentzian components.

amplitude
Vector of amplitudes of the peaks (a.u.)

wavenum
Vector of wavenumbers at which to compute the function.

Value

The value of the pseudo-Voigt function at the given wavenumbers.

References

Thompson, Cox & Hastings (1987) "Rietveld refinement of Debye--Scherrer synchrotron X-ray data from \(Al_2 O_3\)," J. Appl. Crystallogr. 20(2): 79--83, DOI: 10.1107/S0021889887087090

Examples

  Cal_V <- seq(300,400,by=5)
  loc <- c(320,350,375)
  scG <- c(10,5,1)
  scL <- c(3,20,7)
  amp <- c(100,500,200)
  mixedVoigt(loc,scG,scL,amp,Cal_V)

#>  [1]  22.57326  29.15973  47.40780  99.60764 153.07224 132.66580 122.88236
#>  [8] 169.74631 269.21664 414.45077 504.79998 417.16633 276.28023 186.80236
#> [15] 173.25948 272.50413 119.74963  61.83294  41.49159  30.88889  24.24774